Mejora al microservicio de usuario y reseña

+++ frontend/app/src/main/java/com/example/phonecinemaapp/data/remote/dto/ReviewDto.kt
@@ -5,11 +5,12 @@ data class ReviewDto(
     val movieId: Long,
     val userId: Long,
     val userName: String? = null,
-    val rating: Float,
+    val rating: Int,               // ← MAL
     val comment: String,
     val timestamp: String? = null,
     val fotoUsuario: String? = null
 )
+
